Avoid printing an error when the rank for a communicator is negative.
This can happen for MPI_COMM_NULL, there was already a check in the code
for this case however a cast was missing from the test.

+++ src/minfo.c
@@ -431,7 +431,7 @@ int msg_id = 0;
 
 int show_comm (struct process *p, mqs_communicator *comm, int c)
 {
-    if ( comm->local_rank >= 0 )
+    if ( (int)comm->local_rank >= 0 )
 	printf("out: c:%d rank:%d\n",
 	       c,
 	       (int)comm->local_rank);
